+++ src/Resolution.ts
@@ -17,6 +17,8 @@ import {
   SourceDefinition,
   NamehashOptions,
   NamehashOptionsDefault,
+  DnsRecordType,
+  DnsRecord,
   CryptoRecords,
 } from './publicTypes';
 import { nodeHash } from './types';
@@ -25,6 +27,7 @@ import ResolutionError, { ResolutionErrorCode } from './errors/resolutionError';
 import NamingService from './NamingService';
 import { signedInfuraLink } from './utils';
 import { Eip1993Factories } from './utils/Eip1993Factories';
+import DnsUtils from './DnsUtils';
 
 /**
  * Blockchain domain Resolution library - Resolution.
@@ -486,13 +489,30 @@ export default class Resolution {
     return await this.getNamingMethodOrThrow(domain).allRecords(domain);
   }
 
+  async dns(domain: string, types: DnsRecordType[]): Promise<DnsRecord[]> {
+    const dnsUtils = new DnsUtils();
+    domain = this.prepareDomain(domain);
+    const method = this.getNamingMethodOrThrow(domain);
+    const dnsRecordKeys = this.getDnsRecordKeys(types);
+    const blockchainData = await method.records(domain, dnsRecordKeys);
+    return dnsUtils.toList(blockchainData);
+  }
+
+  private getDnsRecordKeys(types: DnsRecordType[]): string[] {
+    const records = ['dns.ttl'];
+    types.forEach(type => {
+      records.push(`dns.${type}`);
+      records.push(`dns.${type}.ttl`);
+    });
+    return records;
+  }
+
   private async getPreferableNewRecord(domain: string, newRecord: string, oldRecord: string): Promise<string> {
-    const records = await this.records(domain, [newRecord, oldRecord]);
+    const records = await this.records(domain, [newRecord, oldRecord]) as Record<string, string>;
     return NamingService.ensureRecordPresence(domain, newRecord, records[newRecord] || records[oldRecord]);
   }
 
   private getNamingMethod(domain: string): NamingService | undefined {
-    domain = this.prepareDomain(domain);
     return this.getResolutionMethods().find(method =>
       method.isSupportedDomain(domain),
     );
@@ -506,7 +526,6 @@ export default class Resolution {
   }
 
   private getNamingMethodOrThrow(domain: string): NamingService {
-    domain = this.prepareDomain(domain);
     const method = this.getNamingMethod(domain);
     if (!method) {
       throw new ResolutionError(ResolutionErrorCode.UnsupportedDomain, {