+++ src/index.ts
@@ -1,8 +1,13 @@
 export { Resolution, Resolution as default } from './Resolution';
 export * from './publicTypes';
+export { default as DnsUtils } from './DnsUtils';
 export { ResolutionError, ResolutionErrorCode } from './errors/resolutionError';
 export {
   ConfigurationError,
   ConfigurationErrorCode,
 } from './errors/configurationError';
+export {
+  DnsRecordsError,
+  DnsRecordsErrorCode
+} from './errors/dnsRecordsError';
 export { Eip1993Factories } from './utils/Eip1993Factories';