+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * Add `Resolution#records` method to query multiple records #96
 * Add formatting options for `Resolution#namehash` and `Resolution#childhash` #91
+* Add `Resolution#dns` method to query dns records from blockchain #99
+* Add DnsUtils as a helper class to convert from CryptoRecord type to DnsRecord and vice-versa #99
 * Plug-in network config from dot-crypto library #101
 
 ## 1.8.3